Digital Readout for iGage scales on the Raspberry Pi
This is the completed first fully working version of PiDRO.  This
project reads digital scales mounted on a milling machine and/or lathe.
The data is displayed on a large computer screen.  Inch or mm display is
provided along with a variety of ways to zero the scales.  The scales
are inexpensive.  This work may aid others in adding scales to their
projects.